Added option @ MoveDirection, Renamed folder, Changed PlayerEntity class.
