+++ src/crosstheborder/lib/Player.java
@@ -1,14 +1,15 @@
-package crossTheBorder.lib;
+package crosstheborder.lib;
 
 /**
  * Created by Oscar on 26-Sep-16.
  * player is the super class for Trump and PlayerEntity
  */
 public abstract class Player {
+
     private String name;
 
-
     public Player(String name) {
+
         this.name = name;
     }
 }
