+++ src/crosstheborder/lib/ability/Crawler.java
@@ -1,4 +1,4 @@
-package crossTheBorder.lib.ability;
+package crosstheborder.lib.ability;
 
 /**
  * Created by Oscar on 26-Sep-16.
