+++ src/crosstheborder/lib/enums/MoveDirection.java
@@ -1,9 +1,9 @@
-package crossTheBorder.lib.enums;
+package crosstheborder.lib.enums;
 
 /**
  * Created by Maurice on 9/26/2016.
  * Class created to tell which directions to player moves to with the hotkeys.
  */
 public enum MoveDirection {
-    UP, DOWN, LEFT, RIGHT
+    UP, DOWN, LEFT, RIGHT, NONE
 }
