+++ src/crosstheborder/lib/player/entity/Mexican.java
@@ -1,4 +1,4 @@
-package crossTheBorder.lib.player.entity;
+package crosstheborder.lib.player.entity;
 
 /**
  * Created by Oscar on 26-Sep-16.
